backend/requirements.txt

flask
flask-sqlalchemy
flask-login
alembic
Flask-Cors
antlr4-python3-runtime
pytest
PyJWT
